Fix lookup methods for UUID pks
The previous return filter tried to filter the values against the original ids list.

When using UUIDs as primary keys, the input ids are of type str(), but
the keys of the dictionary returned by in_bulk are instances of
uuid.UUID.

Another possible fix could be to convert the keys from the dictionary,
before doing the filtering, but since I do not see the point of the
filtering, I chose to use the .values() method instead.