fix: #174 support did add popup in Django 1.10
This changes the way the hook is applied and handles < 1.8 in a simpler way.

Django 1.8 + does call input.trigger('change') after setting the pk,
but Django admin may be using a different jQuery than ajax selects
so event handlers are completely separated. So it is not possible to hook into that.