fix #307 Add new via popup broken
Later versions of Django are appending sequence numbers: __1 to the window name.
These need to be stripped off to find the html id of the input that triggered
the pop up window.

If you open multiple windows for different ajax-select inputs of the same field
then you would get sequences like members__id__2 and it would not work.